fix: throw non-MODULE_NOT_FOUND errors
If there's an error in a file that's not actually a MODULE_NOT_FOUND error, i.e. the file itself has indeed been found but there was an error parsing it, like perhaps an undefined variable being used or a syntax error, then it must throw that error and not mislead the user into thinking that the module/file was somehow not found.